Added a restart function
It sort of works but have to change it so that only the enemies trigger the reload of the scene because right now  the scene loads with any collision that players does.
also i guess i forgot to push my changes during class so the dialogue didnt change. so dont change anything dialogue related please.